Extract CustomFields::FieldDefinition and split file
