+++ app/models/concerns/custom_fields.rb
@@ -46,65 +46,4 @@ module CustomFields
       end
     end
   end
-
-  class CustomField
-    include ActiveModel::Model
-
-    class << self
-      def add_field(key, field_config)
-        attr_accessor key
-        private "#{key}="
-        fields << OpenStruct.new(field_config.merge(key: key))
-        set_type(key, field_config['type']) if field_config['type']
-        set_validation(key, field_config['validations']) if field_config['validations']
-      end
-
-      def fields
-        @fields ||= []
-      end
-
-      def clean(hash)
-        new(hash).as_json.reject { |_, v| v.blank? }.presence
-      end
-
-      private
-
-      def set_type(name, type)
-        case type
-        when 'number'
-          set_validation name, numericality: true
-        end
-      end
-
-      def set_validation(name, validations)
-        validates name, validations
-      end
-    end
-
-    def initialize(*)
-      # accessors need to be public for creation, but are made public
-      # afterwards to try and enforce immutability
-      accessor_methods = fields.map { |f| "#{f.key}=" }
-      self.class.send :public, *accessor_methods
-      super
-      fields.each { |field| send("#{field.key}=", send(field.key) || field.default) }
-      self.class.send :private, *accessor_methods
-    end
-
-    def each(&block)
-      self.class.fields.each(&block)
-    end
-
-    def as_json(*)
-      fields.each_with_object({}) do |field, hash|
-        hash[field.key] = public_send(field.key)
-      end
-    end
-
-    private
-
-    def fields
-      self.class.fields
-    end
-  end
 end
